Use multistage builds to avoid size increases caused by ADD
ADD directive is very handy because it allows to handle both URLs and
local tarballs with the same command, but it increases the size of the
image by the size of the data pulled. This can be mitigated by using
multistage build where stage 0 pulls the tarball, unpacks it and
performs AG installation and stage 1 copies the resulting directory
into a clean base image and proceeds with building the final image.

Change-Id: I44342ccc47dce3c13ab0291d8a94eda960c3455b