Replace recursive chown with find/chown to avoid CoW expansion
Recursive chown at the beginning of the entrypoint is necessary to
ensure whole /agraph directory is owned by agraph:agraph for AG to
operate correctly, but it causes copy-on-write expansion and slows
down the container startup. The new approach only applies chown to
files in the /agraph/data and /agraph/etc directories (mountpoints
of the data and config volumes, respectively) that are not already
owned by agraph:agraph.

Change-Id: Id3542970950a66c3a86fec312b60a5377fc08666